+++ test/Unit/Core/Utilities/LoadCalledCounterTest.ts
@@ -1,39 +1,32 @@
 import 'mocha';
 import * as sinon from 'sinon';
 import { HttpKafka, KafkaCommonObjectType } from 'Core/Utilities/HttpKafka';
-import { LoadCalledCounter } from 'Core/Utilities/LoadCalledCounter';
-import { ABGroup } from 'Core/Models/ABGroup';
+import { LoadCalledCounter, ILoadCalledCounterParams } from 'Core/Utilities/LoadCalledCounter';
 
 describe('LoadCalledCounterTest', () => {
     let sandbox: sinon.SinonSandbox;
     let httpKafkaStub: sinon.SinonStub;
+    const tsDateNow = 1234;
 
     const tests: {
-        kafkaObject: {
-            gameId: string;
-            placementId: string;
-            country: string;
-            count: number;
-            abGroup: ABGroup;
-            organizationId: string;
-            ts: number;
-         };
+        params: ILoadCalledCounterParams;
     }[] = [{
-        kafkaObject: {
+        params: {
             gameId: '1234',
             placementId: 'rewardedVideo',
             count: 1,
             country: 'US',
             abGroup: 99,
             organizationId: 'scottsgames-inc',
-            ts: 1234
+            sdkVersion: 3200,
+            gamerToken: 'secretsecret'
         }
     }];
 
     beforeEach(() => {
         sandbox = sinon.createSandbox();
         httpKafkaStub = sandbox.stub(HttpKafka, 'sendEvent').resolves();
-        sandbox.stub(Date, 'now').returns(1234);
+        sandbox.stub(Date, 'now').returns(tsDateNow);
     });
 
     afterEach(() => {
@@ -43,9 +36,12 @@ describe('LoadCalledCounterTest', () => {
     describe('Sending with correct fields', () => {
         tests.forEach((t) => {
             it(`should send the correct payload"`, () => {
-                const x = t.kafkaObject;
-                LoadCalledCounter.report(x.gameId, x.placementId, x.country, x.count, x.abGroup, x.organizationId);
-                sinon.assert.calledWith(httpKafkaStub, 'ads.load.counting', KafkaCommonObjectType.EMPTY, t.kafkaObject);
+                LoadCalledCounter.report(t.params);
+                sinon.assert.calledWith(httpKafkaStub, 'ads.load.counting', KafkaCommonObjectType.EMPTY, {
+                    ...
+                    t.params,
+                    ts: tsDateNow
+                });
             });
         });
     });